Improve layout, add syscalls, split up files more, etc. improvements

+++ arch/generic/src/stdlib/syscall.c
@@ -1,6 +1,6 @@
 #include <stdlib.h>
 
-intptr_t syscall(long syscall_num, ...)
+intptr_t syscall(intptr_t syscall_num, ...)
 {
 	(void)syscall_num;
 	/* errno = ENOSYS; */
